feat: #WB-3003, reaction choice component
