+++ packages/bootstrap/scss/components/_index.scss
@@ -24,6 +24,7 @@
 @forward "loading-screen";
 @forward "modal";
 @forward "popover";
+@forward "reaction-choice";
 @forward "reaction-summary";
 @forward "select-list";
 @forward "spinner";
