+++ packages/react/ui/src/components/Audience/index.ts
@@ -10,7 +10,11 @@ export * from "./ViewsModal";
 export { default as ReactionSummary } from "./ReactionSummary";
 export * from "./ReactionSummary";
 
-// Component "reactions modal"
+// Component "reactions choice"
+export { default as ReactionChoice } from "./ReactionChoice";
+export * from "./ReactionChoice";
+
+// Component "reactions modal" for details
 export { default as ReactionModal } from "./ReactionModal";
 export * from "./ReactionModal";
